22.05 (#102)
* Rename models.bco to models.BCO
Changes to be committed:
	modified:   bco_api/api/admin.py
	modified:   bco_api/api/models.py
	modified:   bco_api/api/scripts/method_specific/GET_draft_object_by_id.py
	modified:   bco_api/api/scripts/method_specific/GET_published_object_by_id.py
	modified:   bco_api/api/scripts/method_specific/GET_published_object_by_id_with_version.py
	modified:   bco_api/api/scripts/method_specific/POST_api_accounts_new.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_create.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_modify.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_permissions.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_permissions_set.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_publish.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_read.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_token.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_publish.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_published.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_search.py
	modified:   bco_api/api/scripts/utilities/DbUtils.py
	modified:   bco_api/api/signals.py

* Move Groups and Prefix to separate model files

Also added tests for Prefix, Group, and BCO creation
Changes to be committed:
	deleted:    bco_api/api/.vscode/settings.json
	modified:   bco_api/api/admin.py
	new file:   bco_api/api/model/__init__.py
	renamed:    bco_api/api/groups.py -> bco_api/api/model/groups.py
	new file:   bco_api/api/model/prefix.py
	modified:   bco_api/api/models.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_create.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_publish.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_search.py
	deleted:    bco_api/api/scripts/method_specific/POST_api_prefixes_create.py
	deleted:    bco_api/api/scripts/method_specific/POST_api_prefixes_delete.py
	deleted:    bco_api/api/scripts/method_specific/POST_api_prefixes_modify.py
	deleted:    bco_api/api/scripts/method_specific/POST_api_prefixes_permissions_set.py
	deleted:    bco_api/api/scripts/method_specific/POST_api_prefixes_token.py
	deleted:    bco_api/api/scripts/method_specific/POST_api_prefixes_token_flat.py
	modified:   bco_api/api/scripts/utilities/DbUtils.py
	modified:   bco_api/api/scripts/utilities/UserUtils.py
	deleted:    bco_api/api/tests.py
	new file:   bco_api/api/tests/test_bcos.json
	new file:   bco_api/api/tests/test_model_bco.py
	new file:   bco_api/api/tests/test_model_groups.py
	new file:   bco_api/api/tests/test_model_prefix.py
	deleted:    bco_api/api/tests/test_models.py
	modified:   bco_api/api/views.py
	modified:   requirements.txt

* Clean up imports
Changes to be committed:
	modified:   ../.gitignore
	modified:   api/model/groups.py
	modified:   api/model/prefix.py
	modified:   api/models.py
	modified:   api/tests/test_model_groups.py
	modified:   api/tests/test_model_prefix.py
	modified:   api/views.py

* Add dev db
Changes to be committed:
	new file:   admin_only/db.sqlite3.dev

* Create codeql-analysis.yml (#101)

Adding in the basic code quality analysis scan.
