--- bco_api/api/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "python.linting.pylintEnabled": true,
-    "python.linting.enabled": true
-}
