+++ bco_api/api/scripts/method_specific/GET_draft_object_by_id.py
@@ -6,7 +6,7 @@ see if the requestor has permissions
 for it.
 """
 
-from api.models import bco
+from api.models import BCO
 from api.scripts.utilities import UserUtils
 from rest_framework import status
 from rest_framework.response import Response
@@ -28,7 +28,7 @@ def GET_draft_object_by_id(do_id, request):
         it is returned. If not the response is HTTP_403_FORBIDDEN.
     """
 
-    filtered = bco.objects.filter(object_id__regex=rf'(.*?)/{do_id}', state="DRAFT")
+    filtered = BCO.objects.filter(object_id__regex=rf'(.*?)/{do_id}', state="DRAFT")
 
     if filtered.exists():
         if len(filtered) > 1:
@@ -40,11 +40,11 @@ def GET_draft_object_by_id(do_id, request):
             )
         # Get the requestor's info.
         usr_info = UserUtils.UserUtils().user_from_request(request=request)
-        user_objects = get_objects_for_user(user=usr_info, perms=[], klass=bco, any_perm=True)
+        user_objects = get_objects_for_user(user=usr_info, perms=[], klass=BCO, any_perm=True)
 
         # Does the requestor have permissions for the object?
         full_object_id = filtered.values_list('object_id', flat=True)[0]
-        objected = bco.objects.get(object_id=full_object_id)
+        objected = BCO.objects.get(object_id=full_object_id)
         if objected in user_objects:
             return Response(data=objected.contents, status=status.HTTP_200_OK)
         # Insufficient permissions.
