+++ bco_api/api/scripts/method_specific/POST_api_accounts_new.py
@@ -110,7 +110,7 @@ def POST_api_accounts_new(request):
             print('activation_link', activation_link)
             # print('ERROR: ', e)
             # TODO: Should handle when the send_mail function fails?
-            return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR, data={"message": "Not able to send authentication email: {}".format(e)})
+            #return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR, data={"message": "Not able to send authentication email: {}".format(e)})
         return Response(status=status.HTTP_201_CREATED)
 
     else:
