+++ bco_api/api/scripts/method_specific/POST_api_objects_drafts_permissions.py
@@ -1,5 +1,5 @@
 # BCO model
-from ...models import bco
+from api.models import BCO
 
 # For getting objects out of the database.
 from ..utilities import DbUtils
@@ -55,7 +55,7 @@ def POST_api_objects_drafts_permissions(incoming):
         # Does the requestor have view permissions for
         # the *prefix*?
         if 'view_' + standardized in px_perms:
-            print('bulk_request', list(bco.objects.filter(object_id=creation_object['object_id'])))
+            print('bulk_request', list(BCO.objects.filter(object_id=creation_object['object_id'])))
             # The requestor has change view for
             # the prefix, but do they have object-level
             # view permissions?
@@ -66,8 +66,8 @@ def POST_api_objects_drafts_permissions(incoming):
             # group that has object-level view permissions.
 
             # To check these options, we need the actual object.
-            if bco.objects.filter(object_id=creation_object['object_id']).exists():
-                objected = bco.objects.get(object_id=creation_object['object_id'])
+            if BCO.objects.filter(object_id=creation_object['object_id']).exists():
+                objected = BCO.objects.get(object_id=creation_object['object_id'])
 
                 # We don't care where the view permission comes from,
                 # be it a User permission or a Group permission.
