+++ bco_api/api/scripts/method_specific/POST_api_objects_drafts_read.py
@@ -1,5 +1,5 @@
 # BCO model
-from ...models import bco
+from ...models import BCO
 
 # For getting objects out of the database.
 from ..utilities import DbUtils
@@ -66,8 +66,8 @@ def POST_api_objects_drafts_read(incoming):
             # group that has object-level view permissions.
 
             # To check these options, we need the actual object.
-            if bco.objects.filter(object_id=read_object['object_id']).exists():
-                objected = bco.objects.get(
+            if BCO.objects.filter(object_id=read_object['object_id']).exists():
+                objected = BCO.objects.get(
                         object_id=read_object['object_id']
                         )
