+++ bco_api/api/scripts/method_specific/POST_api_objects_drafts_token.py
@@ -8,7 +8,7 @@ import re
 # Concatenating QuerySets
 from itertools import chain
 from typing import Optional, Tuple
-from api.models import bco
+from api.models import BCO
 from api.scripts.utilities import UserUtils
 # Object-level permissions
 from guardian.shortcuts import get_objects_for_user
@@ -121,10 +121,10 @@ def POST_api_objects_drafts_token(rqst, internal=False):
 
     # Use an empty list of perms to get ANY perm.
     # Source: https://stackoverflow.com/a/24980558
-    user_objects = get_objects_for_user(user=user_info, perms=[], klass=bco, any_perm=True)
+    user_objects = get_objects_for_user(user=user_info, perms=[], klass=BCO, any_perm=True)
 
     # Now get all objects under these prefixes.
-    prefix_objects = bco.objects.filter(prefix__in=user_prefixes, state='DRAFT')
+    prefix_objects = BCO.objects.filter(prefix__in=user_prefixes, state='DRAFT')
 
     # Assume all the values are supposed to be returned.
     # Source: https://stackoverflow.com/a/51733590
@@ -178,7 +178,7 @@ def POST_api_objects_drafts_token(rqst, internal=False):
         # add in the published objects.
         # TODO: This needs to only return the most recent PUBLISHED objects not all of the versions
 
-        published = bco.objects.filter(state='PUBLISHED').values()
+        published = BCO.objects.filter(state='PUBLISHED').values()
         # unique_published = []
         unique_published = set()
