+++ requirements.txt
@@ -1,5 +1,6 @@
 appdirs==1.4.3
 asgiref==3.3.4
+astroid==2.9.3
 attrs==20.3.0
 CacheControl==0.12.6
 certifi==2019.11.28
@@ -9,6 +10,7 @@ colorama==0.4.3
 contextlib2==0.6.0
 coreapi==2.3.3
 coreschema==0.0.4
+coverage==6.3.2
 distlib==0.3.0
 distro==1.4.0
 Django==3.2.10
@@ -27,18 +29,25 @@ html5lib==1.0.1
 idna==2.8
 inflection==0.5.1
 ipaddr==2.2.0
+isort==5.10.1
 itypes==1.2.0
 Jinja2==3.0.1
 jsonref==0.2
 jsonschema==3.2.0
+lazy-object-proxy==1.7.1
 lockfile==0.12.2
 MarkupSafe==2.0.1
+mccabe==0.6.1
 msgpack==0.6.2
 openapi-codec==1.3.2
 packaging==20.3
 pep517==0.8.2
+platformdirs==2.5.1
 progress==1.5
 PyJWT==1.7.1
+pylint==2.12.2
+pylint-django==2.5.3
+pylint-plugin-utils==0.7
 pyparsing==2.4.6
 pyrsistent==0.17.3
 python-dateutil==2.8.1
@@ -53,7 +62,9 @@ simplejson==3.17.5
 six==1.14.0
 sqlparse==0.4.2
 toml==0.10.2
+typing_extensions==4.1.1
 uritemplate==3.0.1
 urllib3==1.26.5
 webencodings==0.5.1
+wrapt==1.13.3
 xmltodict==0.12.0
